ai_agent_project/requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
flake8>=6.1.0 